+++ monas-account/src/application_service.rs
@@ -1,0 +1,1 @@
+pub(crate) mod account_service;
