+++ monas-account/src/domain.rs
@@ -1,1 +1,1 @@
-mod account;
+pub(crate) mod account;
