+++ monas-account/src/domain/account.rs
@@ -1,8 +1,4 @@
-use secret::key_pair::{KeyPair, KeyType};
-use crate::domain::account::secret::key_pair::k256_key_pair::K256KeyPair;
-use crate::domain::account::secret::key_pair::p256_key_pair::P256KeyPair;
-
-pub(crate) mod secret;
+use crate::infrastructure::key_pair::KeyPair;
 
 pub struct Account {
     key_pair: KeyPair,
@@ -16,12 +12,18 @@ pub enum AccountError {
 
 
 impl Account {
+    pub(crate) fn init(key_pair: KeyPair) -> Self {
+        Account {
+            key_pair,
+            deleted: false,
+        }
+    }
 
-    pub fn regenerate_keypair(&mut self) -> Result<(), AccountError> {
+    pub fn regenerate_keypair(&mut self, key_pair: KeyPair) -> Result<(), AccountError> {
         if self.deleted {
             return Err(AccountError::AccountAlreadyDeleted);
         }
-       self.key_pair = self.key_pair.generate();
+        self.key_pair = key_pair;
         Ok(())
     }
 
@@ -42,40 +44,14 @@ impl Account {
     }
 }
 
-struct AccountFactory;
-
-impl AccountFactory {
-    pub fn create(key_type: KeyType) -> Account {
-        let key_pair = match key_type {
-            KeyType::K256 => KeyPair::K256KeyPair(K256KeyPair::generate()),
-            KeyType::P256 => KeyPair::P256KeyPair(P256KeyPair::generate())
-        };
-        Account {
-            key_pair,
-            deleted: false,
-        }
-    }
-}
-
 #[cfg(test)]
 mod account_tests {
-    use crate::domain::account::secret::key_pair::KeyType::K256;
+    use crate::infrastructure::key_pair::KeyType::K256;
     use super::*;
-    #[test]
-    fn create_account() {
-        let account = AccountFactory::create(K256);
-        let is_created_key_type =
-            match &account.key_pair {
-                KeyPair::K256KeyPair(k256) => true,
-                _ => false,
-            }.clone();
-        assert_eq!(is_created_key_type, true);
-        assert_eq!(account.is_deleted(), false);
-    }
 
     #[test]
     fn regenerate_key_pair() {
-        let mut account = AccountFactory::create(K256);
+        let mut account = Account::init(KeyPair::generate(K256));
 
         let key_pair_before =
             match &account.key_pair {
@@ -83,7 +59,7 @@ mod account_tests {
                 _ => panic!("unexpected public key detected"),
             }.clone();
 
-        account.regenerate_keypair().unwrap();
+        account.regenerate_keypair(KeyPair::generate(K256)).unwrap();
 
         let key_pair_after = match &account.key_pair {
             KeyPair::K256KeyPair(k256) => k256.secret_key(),
@@ -97,14 +73,15 @@ mod account_tests {
 
     #[test]
     fn throw_error_regenerate_key_pair_when_account_was_deleted() {
-        let mut account = AccountFactory::create(K256);
+        let mut account = Account::init(KeyPair::generate(K256));
+
         account.delete().unwrap();
         assert!(account.is_deleted());
-        assert_eq!(account.regenerate_keypair(), Err(AccountError::AccountAlreadyDeleted));
+        assert_eq!(account.regenerate_keypair(KeyPair::generate(K256)), Err(AccountError::AccountAlreadyDeleted));
     }
 
     fn delete_account() {
-        let mut account = AccountFactory::create(K256);
+        let mut account = Account::init(KeyPair::generate(K256));
         account.delete().unwrap();
         assert!(account.is_deleted());
     }
