--- monas-account/src/domain/account/secret.rs
@@ -1,1 +1,0 @@
-pub mod key_pair;
