+++ monas-account/src/infrastructure.rs
@@ -1,0 +1,1 @@
+pub(crate) mod key_pair;
