+++ monas-account/src/lib.rs
@@ -1,4 +1,7 @@
 mod domain;
+mod infrastructure;
+
+mod application_service;
 
 pub fn add(left: usize, right: usize) -> usize {
     left + right
